moved ibip to simulations, added deprecation warnings
