+++ porespy/simulations/__init__.py
@@ -18,3 +18,5 @@ This module contains routines for performing simulations directly on images.
 
 from ._drainage import *
 from ._dns import *
+from ._ibip import ibip
+from ._ibip_gpu import ibip_gpu
